+++ package.json
@@ -19,6 +19,7 @@
     "emotion": "^11.0.0",
     "marked": "^4.0.8",
     "next": "12.0.7",
+    "qs": "^6.10.3",
     "react": "17.0.2",
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "17.0.2",
@@ -26,6 +27,7 @@
   },
   "devDependencies": {
     "@types/node": "16.11.12",
+    "@types/qs": "^6.9.7",
     "@types/react": "17.0.37",
     "@types/styled-components": "^5.1.17",
     "babel-plugin-styled-components": "^2.0.2",
