+++ src/components/MyPage/Menus/Content.tsx
@@ -1,57 +1,48 @@
 import styled from "@emotion/styled"
-import { API_ENDPOINT, MEDIA_QUERY_END_POINT } from "../../../constants"
+import { MEDIA_QUERY_END_POINT } from "../../../constants"
 import SearchIcon from "@mui/icons-material/Search"
-import { MyCard } from "../MyCard"
-import { useContext, useEffect, useRef, useState } from "react"
+import { useContext, useRef, useState } from "react"
 import { ThemeContext } from "../../../pages/_app"
 import { ThemeProps } from "../../../types/Theme"
-import CardLoading from "../../Common/CardLoading"
-import { fetcher } from "../../../utils/fetcher"
-import useSWRInfinite from "swr/infinite"
+import { ContentData } from "./ContentData"
+import { useData } from "../../../hooks/useData"
+import qs from "qs"
 
 interface ContentProps {
     username: string | string[] | undefined
 }
-const PAGE_SIZE = 2
+
 export const Content = ({ username }: ContentProps) => {
     const { theme } = useContext(ThemeContext)
+    const [tag, setTag] = useState("")
+    const tagData = useRef([])
 
-    const [target, setTarget] = useState<HTMLElement | null | undefined>(null)
-
-    const getKey = (pageIndex: number, previousPageData: any) => {
-        if (previousPageData && !previousPageData.data) return null
-        return `${API_ENDPOINT}/posts?pagination[page]=${pageIndex}&pagination[pageSize]=${PAGE_SIZE}`
-    }
-
-    const { data, size, setSize, error, isValidating } = useSWRInfinite(
-        getKey,
-        fetcher
+    const query = qs.stringify(
+        {
+            populate: {
+                posts: {
+                    populate: ["userid"],
+                },
+            },
+        },
+        {
+            encodeValuesOnly: true,
+        }
     )
 
-    const isLoadingInitialData = !data && !error
-    const isLoadingMore =
-        isLoadingInitialData ||
-        (size > 0 && data && typeof data[size - 1] === "undefined")
+    const { data, error, isValidating } = useData("hashtags", query)
 
-    const isEmpty = data?.[0]?.length === 0
-    const isReachingEnd =
-        isEmpty || (data && data[data.length - 1]?.length < PAGE_SIZE)
-
-    const onIntersect: IntersectionObserverCallback = ([entry], observer) => {
-        if (entry.isIntersecting) {
-            setSize((prev) => prev + 1)
-            if (isReachingEnd) observer.unobserve(entry.target)
-        }
+    if (!isValidating) {
+        tagData.current = data.data.filter((e: any) =>
+            e.attributes.posts.data.some(
+                (post: any) =>
+                    post.attributes.userid.data?.attributes.nickname ===
+                    username
+            )
+        )
     }
 
-    useEffect(() => {
-        if (!target || isReachingEnd) return
-        const observer = new IntersectionObserver(onIntersect, {
-            threshold: 0.4,
-        })
-        observer.observe(target)
-        return () => observer && observer.disconnect()
-    }, [target])
+    const handleTag = () => {}
 
     return (
         <ContentContainer>
@@ -68,7 +59,27 @@ export const Content = ({ username }: ContentProps) => {
                             전체보기<span>(50)</span>
                         </a>
                     </li>
-                    {/* 태그 데이터 받아서 추가 */}
+                    {tagData.current.map((tag: any) => {
+                        return (
+                            <li key={tag.attributes.id}>
+                                <a>
+                                    {tag.attributes.name}
+                                    <span>
+                                        (
+                                        {
+                                            tag.attributes.posts.data.filter(
+                                                (e: any) =>
+                                                    e.attributes.userid.data
+                                                        ?.attributes
+                                                        .nickname === username
+                                            ).length
+                                        }
+                                        )
+                                    </span>
+                                </a>
+                            </li>
+                        )
+                    })}
                 </ul>
             </SmallTaglist>
             <LargeTaglist theme={theme}>
@@ -78,31 +89,30 @@ export const Content = ({ username }: ContentProps) => {
                         <a>전체보기</a>
                         <span>(50)</span>
                     </li>
-                    {/* 태그 데이터 받아서 추가 */}
+                    {tagData.current.map((tag: any) => {
+                        return (
+                            <li key={tag.attributes.id}>
+                                <a>
+                                    {tag.attributes.name}
+                                    <span>
+                                        (
+                                        {
+                                            tag.attributes.posts.data.filter(
+                                                (e: any) =>
+                                                    e.attributes.userid.data
+                                                        ?.attributes
+                                                        .nickname === username
+                                            ).length
+                                        }
+                                        )
+                                    </span>
+                                </a>
+                            </li>
+                        )
+                    })}
                 </ul>
             </LargeTaglist>
-            <section>
-                {data &&
-                    data
-                        .filter((e, i) => i != 0)
-                        .map((loaded) => {
-                            return loaded.data.map((e: any, i: number) => (
-                                <MyCard
-                                    key={i}
-                                    imageUrl={e.attributes.imageUrl}
-                                    title={e.attributes.title}
-                                    contents={e.attributes.contents}
-                                    tag={e.attributes.tag?.tags}
-                                    date={e.attributes.publichedAt}
-                                    // comment={e.attributes.comment}
-                                    username={username}
-                                />
-                            ))
-                        })}
-            </section>
-            <TargetElement ref={setTarget}>
-                {isLoadingMore && !isReachingEnd && <CardLoading />}
-            </TargetElement>
+            <ContentData username={username} />
         </ContentContainer>
     )
 }
@@ -218,8 +228,4 @@ const SmallTaglist = styled.section<ThemeProps>`
             margin-left: 8px;
         }
     }
-`
-const TargetElement = styled.article`
-    width: 100%;
-    height: 100px;
 `
