+++ src/hooks/useData.ts
@@ -2,6 +2,6 @@ import useSWR from "swr"
 import { fetcher } from "../utils/fetcher"
 import { API_ENDPOINT } from "../constants"
 
-export const useData = (path: string) => {
-  return useSWR(`${API_ENDPOINT}/${path}`, fetcher)
+export const useData = (path: string, query: string = "") => {
+    return useSWR(`${API_ENDPOINT}/${path}?${query}`, fetcher)
 }
